+++ CHANGES.md
@@ -1,6 +1,12 @@
 # Release notes
 All notable changes to this project will be documented in this file.
 
+## 0.2.0
+- Migrate data + add rarity of cards to battle stores 
+- Add filter and sort options to home page
+- Add top 5 cards to home page sorted
+- Complete table now behind accordion on the home page
+- Update on how land price is calculated when not a perfect match is found take closed match 
 
 ## 0.1.0
  - Initial release